+++ src/SonsOfPHP/Contract/Cqrs/CommandMessageHandlerInterface.php
@@ -7,6 +7,4 @@ namespace SonsOfPHP\Contract\Cqrs;
 /**
  * @author Joshua Estes <joshua@sonsofphp.com>
  */
-interface CommandMessageHandlerInterface extends MessageHandlerInterface
-{
-}
+interface CommandMessageHandlerInterface extends MessageHandlerInterface {}