+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Examples:
 * [PR #70](https://github.com/SonsOfPHP/sonsofphp/pull/70) Added new Core contract
 * [PR #112](https://github.com/SonsOfPHP/sonsofphp/pull/112) [Cache] Added new component
 * [PR #119](https://github.com/SonsOfPHP/sonsofphp/pull/119) [HttpMessage] Added `withQueryParams` and `withQueryParam` to `Uri`
+* [PR #120](https://github.com/SonsOfPHP/sonsofphp/pull/120) [Logger] New Component
 
 ## [0.3.8]